requirements.txt
streamlit
python-dotenv
youtube-transcript-api
langchain-groq
langchain
pymupdf
Pillow
google-generativeai
pytubefix
pydub
openai